Adding my first app
I've created an app with 3 pages. at first I created a view in mysite folder and worked with urls.py in mysite folder. then I created an app and then created urls in the app. after that i included the urls in mysite folder and then I deleted the views in mysite and transferred it to the app I ceated